multi-file symbol loading and more
