+++ LICENSE
@@ -1,5 +1,6 @@
 MIT License
 
+Copyright (c) 2024 Dave Lee
 Copyright (c) 2024 Dániel Bali
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
