+++ index.html
@@ -17,6 +17,7 @@
     <div class="wrapper"><button class="roll">🕹️</button></div>
     <div class="wrapper"><div class="shop"></div></div>
     <div class="wrapper"><div class="inventory"></div></div>
+    <div class="wrapper"><div class="settings"></div></div>
   </div></div>
   <div class="discord">
     <a target="_blank" href="https://discord.gg/wyCDHsuk">
