+++ main.js
@@ -1,113 +1,10 @@
-import {
-  Symbol, Empty,
-  Balloon,
-  Bank,
-  Bell,
-  Bomb,
-  Briefcase,
-  Bubble,
-  Butter,
-  Bug,
-  BullsEye,
-  Champagne,
-  Cherry,
-  Chick,
-  Chicken,
-  Clover,
-  Cocktail,
-  Coin,
-  Corn,
-  CreditCard,
-  CrystalBall,
-  Dancer,
-  Diamond,
-  Dice,
-  Dragon,
-  Drums,
-  Egg,
-  Firefighter,
-  Fox,
-  FreeTurn,
-  Grave,
-  Hole,
-  MagicWand,
-  Mango,
-  MoneyBag,
-  Moon,
-  Multiplier,
-  MusicalNote,
-  Pineapple,
-  Popcorn,
-  Record,
-  Refresh,
-  Rock,
-  Rocket,
-  ShoppingBag,
-  Slots,
-  Snail,
-  Tree,
-  Volcano,
-  Worker,
-} from './symbol.js';
 import * as Util from './util.js'
-
-const makeCatalog = () => [
-  new Balloon(),
-  new Bank(),
-  new Bell(),
-  new Bomb(),
-  new Briefcase(),
-  new Bug(),
-  new BullsEye(),
-  new Butter(),
-  new Champagne(),
-  new Cherry(),
-  new Chick(),
-  new Chicken(),
-  new Clover(),
-  new Cocktail(),
-  new Coin(),
-  new Corn(),
-  new CreditCard(),
-  new CrystalBall(),
-  new Dancer(),
-  new Diamond(),
-  new Dice(),
-  new Dragon(),
-  new Drums(),
-  new Egg(),
-  // new Firefighter(),
-  new Fox(),
-  // new FreeTurn(),
-  // new Grave(),
-  new Hole(),
-  new MagicWand(),
-  new Mango(),
-  new MoneyBag(),
-  new Moon(),
-  new Multiplier(),
-  new Pineapple(),
-  new Record(),
-  new Refresh(),
-  new Rock(),
-  new Rocket(),
-  new ShoppingBag(),
-  new Slots(),
-  new Snail(),
-  new Tree(),
-  // new Volcano(),
-  new Worker(),
-];
-
-const startingSet = () => [
-  new Coin(),
-  new Cherry(),
-  new Cherry(),
-  new Cherry(),
-];
+import { GameSettings } from './game_settings.js'
+import { Catalog } from './catalog.js';
+import { Board } from './board.js';
 
 // Test
-makeCatalog().forEach(s => s.copy());
+// makeCatalog().forEach(s => s.copy());
 let totalTurns = 0;
 
 class Inventory {
@@ -181,7 +78,8 @@ class Inventory {
 }
 
 class Shop {
-  constructor() {
+  constructor(catalog) {
+    this.catalog = catalog;
     this.shopDiv = document.querySelector('.game .shop');
     this.isOpen = false;
     this.refreshCost = 1;
@@ -205,19 +103,11 @@ class Shop {
       return total;
     };
     let luck = 0;
-    luck += checkLuckyItem(Clover.name, 0.01);
-    luck += checkLuckyItem(CrystalBall.name, 0.03);
+    luck += checkLuckyItem('🍀', 0.01);
+    luck += checkLuckyItem('🔮', 0.03);
 
     this.shopDiv.replaceChildren();
-    this.catalog = makeCatalog();
-    const newCatalog = [];
-    while (newCatalog.length < 3) {
-      for (const item of this.catalog) {
-        if (Math.random() < item.rarity + luck) {
-          newCatalog.push(item);
-        }
-      }
-    }
+    const newCatalog = this.catalog.generate_shop(3, luck)
 
     const makeShopItem = (symbol, description, handler, refresh=false) => {
       const shopItemDiv = document.createElement('div');
@@ -299,207 +189,17 @@ class Shop {
 
 }
 
-class Board {
-  constructor() {
-    this.cells = [];
-    this.gridDiv = document.querySelector('.game .grid');
-    this.gridDiv.replaceChildren();
-    for (let i = 0; i < Util.BOARD_SIZE; ++i) {
-      const row = [];
-      const rowDiv = document.createElement('div');
-      rowDiv.classList.add('row');
-      for (let j = 0; j < Util.BOARD_SIZE; ++j) {
-        row.push(new Empty());
-        const cellDiv = document.createElement('div');
-        cellDiv.classList.add('cell');
-        const symbolDiv = document.createElement('div');
-        symbolDiv.classList.add('symbol');
-        symbolDiv.innerText = '⬜';
-        const counterDiv = document.createElement('div');
-        counterDiv.classList.add('symbolCounter');
-        counterDiv.innerText = '';
-        cellDiv.appendChild(symbolDiv);
-        cellDiv.appendChild(counterDiv);
-        rowDiv.appendChild(cellDiv);
-      }
-      this.cells.push(row);
-      this.gridDiv.appendChild(rowDiv);
-    }
-  }
-  getSymbolDiv(x, y) {
-    return this.gridDiv.children[y].children[x].children[0];
-  }
-  updateCounter(game, x, y) {
-    const counterDiv = this.gridDiv.children[y].children[x].children[1];
-    const counter = this.cells[y][x].counter(game);
-    if (counter !== null) {
-      counterDiv.innerText = counter;
-    }
-  }
-  getCounterDiv(x, y) {
-    return this.gridDiv.children[y].children[x].children[1];
-  }
-  async showMoneyEarned(x, y, value) {
-    const moneyDiv = document.createElement('div');
-    moneyDiv.classList.add('moneyEarned');
-    moneyDiv.innerText = `💵${value}`;
-    this.gridDiv.appendChild(moneyDiv);
 
-    await Util.animate(moneyDiv, 'fadeOutMoveDown', 0.3);
-    this.gridDiv.removeChild(moneyDiv);
-  }
-  clearCell(x, y) {
-    this.getCounterDiv(x, y).innerText = '';
-    this.cells[y][x] = new Empty();
-  }
-  async spinDiv(game, x, y, symbol) {
-    await Util.delay(Util.random(600));
-    const div = this.getSymbolDiv(x, y);
-    const counterDiv = this.getCounterDiv(x, y);
-    counterDiv.innerText = '';
-    const randomSymbol = () => {
-      const set = new Set();
-      for (const symbol of Object.values(game.inventory.symbols)) {
-        set.add(symbol.name());
-      }
-      div.innerText = Util.randomChoose([...set]);
-    }
-    await Util.animate(div, 'startSpin', 0.1);
-    for (let i = 0; i < 6; ++i) {
-      randomSymbol();
-      await Util.animate(div, 'spin', 0.12 + i * 0.02);
-    }
-    div.innerText = symbol.name();
-    await Util.animate(div, 'endSpin', 0.3);
-    await Util.animate(div, 'bounce', 0.1);
-    if (symbol.counter(game) != null) {
-      counterDiv.innerText = symbol.counter(game);
-    }
-  }
-  async spinDivOnce(game, x, y) {
-    const div = this.getSymbolDiv(x, y);
-    const counterDiv = this.getCounterDiv(x, y);
-    counterDiv.innerText = '';
-    await Util.animate(div, 'startSpin', 0.1);
-    const symbol = this.cells[y][x];
-    div.innerText = symbol.name();
-    await Util.animate(div, 'endSpin', 0.3);
-    await Util.animate(div, 'bounce', 0.1);
-    if (symbol.counter(game) != null) {
-      counterDiv.innerText = symbol.counter(game);
-    }
-  }
-  async roll(game) {
-    const symbols = [...game.inventory.symbols];
-    const empties = [];
-    for (let i = 0; i < Util.BOARD_SIZE; ++i) {
-      for (let j = 0; j < Util.BOARD_SIZE; ++j) {
-        empties.push([j, i]);
-        this.cells[i][j] = new Empty();
-      }
-    }
-    for (let i = 0; i < Util.BOARD_SIZE * Util.BOARD_SIZE; ++i) {
-      if (symbols.length === 0) {
-        break;
-      }
-      const symbol = Util.randomRemove(symbols)
-      const [x, y] = Util.randomRemove(empties);
-      this.cells[y][x] = symbol;
-    }
-    const tasks = [];
-    for (let i = 0; i < Util.BOARD_SIZE; ++i) {
-      for (let j = 0; j < Util.BOARD_SIZE; ++j) {
-        tasks.push(
-          this.spinDiv(game, j, i, this.cells[i][j]));
-      }
-    }
-    await Promise.all(tasks);
-  }
-  async evaluate(game) {
-    this.forAllCells((cell, x, y) => {
-      cell.turns++;
-    });
-    this.forAllCells((cell, x, y) => {
-      this.updateCounter(game, x, y);
-    });
-    const evaluateRound = async (f) => {
-      const tasks = [];
-      this.forAllCells((cell, x, y) => tasks.push(
-        async () => { 
-          // If the symbol has since been removed from the board, do not evaluate.
-          if (this.cells[y][x] !== cell) {
-            return;
-          }
-          await f(cell, game, x, y);
-        } ));
-      for (const task of tasks) {
-        await task();
-      }
-      this.forAllCells((cell, x, y) => {
-        this.updateCounter(game, x, y);
-      });
-    };
-    await evaluateRound((c, g, x, y) => c.evaluateConsume(g, x, y));
-    await evaluateRound((c, g, x, y) => c.evaluateProduce(g, x, y));
-    await evaluateRound((c, g, x, y) => c.evaluateConsume(g, x, y));
-  }
-  async finalScore(game) {
-    const tasks = [];
-    this.forAllCells((cell, x, y) => {
-      tasks.push(async () => {
-        await cell.finalScore(game, x, y);
-      });
-    })
-    for (const task of tasks) {
-      await task();
-    }
-  }
-  async score(game) {
-    const tasks = [];
-    this.forAllCells((cell, x, y) => {
-      tasks.push(async () => {
-        await cell.score(game, x, y);
-      });
-    })
-    for (const task of tasks) {
-      await task();
-    }
-  }
-  forAllCells(f) {
-    this.cells.forEach((row, y) => {
-      row.forEach((cell, x) => {
-        f(cell, x, y);
-      });
-    });
-  }
-  async addSymbol(game, sym, x, y) {
-    game.inventory.add(sym);
-    if (this.cells[y][x].name() === Hole.name) {
-      const hole = this.cells[y][x];
-      this.cells[y][x] = sym;
-      await this.spinDivOnce(game, x, y);
-      this.cells[y][x] = hole;
-      await this.spinDivOnce(game, x, y);
-    } else {
-      this.cells[y][x] = sym;
-      await this.spinDivOnce(game, x, y);
-    }
-    this.updateCounter(game, x, y);
-  }
-  async removeSymbol(game, x, y) {
-    game.inventory.remove(this.cells[y][x]);
-    this.clearCell(x, y);
-    await Util.animate(this.getSymbolDiv(x, y), 'flip', 0.15);
-    await this.spinDivOnce(game, x, y);
-  }
-}
 
 class Game {
-  constructor() {
-    this.inventory = new Inventory(startingSet());
+  constructor(gameSettings, catalog) {
+    this.gameSettings = gameSettings;
+    this.catalog = catalog;
+    console.log(this.catalog);
+    this.inventory = new Inventory(this.catalog.symbolsFromString(this.gameSettings.startingSet));
     this.inventory.update();
-    this.board = new Board();
-    this.shop = new Shop();
+    this.board = new Board(this.gameSettings, this.catalog);
+    this.shop = new Shop(this.catalog);
     this.rolling = false;
     this.info = document.querySelector('.game .info');
     Util.drawText(this.info, "hi there. press (🕹️) when you are ready to play.");
@@ -583,7 +283,7 @@ class Game {
   }
 }
 
-const load = () => {
+const load = async () => {
   document.querySelector('body').removeEventListener(
     'click', load);
   const template = document.querySelector('.template');
@@ -592,7 +292,10 @@ const load = () => {
   const templateClone = template.cloneNode(true);
   templateClone.classList.remove('hidden');
   gameDiv.appendChild(templateClone.children[0]);
-  const game = new Game();
+  const gameSettings = new GameSettings();
+  const catalog = new Catalog(gameSettings.symbolSources)
+  await catalog.updateSymbols();
+  const game = new Game(gameSettings, catalog);
   return game;
 };
